# V3: Simplify asset and dependency processing