--- a/crates/atlaspack_core/src/lib.rs
+++ b/crates/atlaspack_core/src/lib.rs
@@ -4,4 +4,5 @@ pub mod cache;
 pub mod config_loader;
 pub mod hash;
 pub mod plugin;
+pub mod project_path;
 pub mod types;