## Motivation

Once the rust based asset graph is returned, adding all the nodes to the JavaScript based asset graph is quite slow. These changes reduce the amount of fields that are changed to what is necessary. Both the three-js benchmarks and an internal product build show a ~45-65% improvement on the `getAssetGraph` function with these changes.

| Benchmark  | Before | After | Improvement
| - | - | - | - |
| internal | 11.40s | 4.09s | -7.31s (64.14%) |
| three-js | 1.28s  | 0.68s  | -0.6s (46.88%) |

## Changes

* Mutate asset and dependency values
* Only change what is needed

## Checklist

- [x] Existing or new tests cover this change